requirements.txt
Pillow
opencv-python-headless
numpy
PyWavelets
torch
torchvision
transformers
exifread
folium
streamlit
